+++ R/utils.R
@@ -1,5 +1,6 @@
-col_name <- function(x) {
+col_name <- function(x, default = stop("Please supply column name", call. = FALSE)) {
   if (is.character(x)) return(x)
+  if (identical(x, quote(expr = ))) return(default)
   if (is.name(x)) return(as.character(x))
   if (is.null(x)) return(x)