+++ tests/testthat/test-gather.R
@@ -9,3 +9,10 @@ test_that("gather all columns when ... is empty", {
   expect_equal(nrow(out), 10)
   expect_equal(names(out), c("key", "val"))
 })
+
+test_that("if not supply, key and value default to key and value", {
+  df <- data.frame(x = 1:2)
+  out <- gather(df)
+  expect_equal(nrow(out), 2)
+  expect_equal(names(out), c("key", "value"))
+})